requirements.txt
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
deep-translator>=1.11.0
google-generativeai>=0.3.0
python-dateutil>=2.8.0
lxml>=4.9.0
